pyproject.toml
[project]
name = "highschool-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-core>=0.3.75",
    "langchain-groq>=0.3.7",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.7",
    "langsmith>=0.4.25",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
    "sympy>=1.14.0",
]
